Modefy according to interactive task requirements
